#1710 move to package singleton
